Add detailed README and enhance HuggingFace model client
Introduce a comprehensive README detailing the insurance policy analysis pipeline architecture, methodology, and usage. Refactor the HuggingFace model client to include Persona and JSON extractors, improving modularity, and remove redundant JSON extraction methods for cleaner implementation.
